address pr comments
