+++ vscode/src/completions/context/context-mixer.ts
@@ -66,6 +66,11 @@ export interface GetContextResult {
     contextLoggingSnippets: AutocompleteContextSnippet[]
 }
 
+export interface ContextMixerOptions {
+    strategyFactory: ContextStrategyFactory
+    dataCollectionEnabled?: boolean
+}
+
 /**
  * The context mixer is responsible for combining multiple context retrieval strategies into a
  * single proposed context list.
@@ -78,11 +83,10 @@ export interface GetContextResult {
 export class ContextMixer implements vscode.Disposable {
     private disposables: vscode.Disposable[] = []
     private contextDataCollector: ContextRetrieverDataCollection | null = null
+    private strategyFactory: ContextStrategyFactory
 
-    constructor(
-        private strategyFactory: ContextStrategyFactory,
-        dataCollectionEnabled = false
-    ) {
+    constructor({ strategyFactory, dataCollectionEnabled = false }: ContextMixerOptions) {
+        this.strategyFactory = strategyFactory
         if (dataCollectionEnabled) {
             this.contextDataCollector = new ContextRetrieverDataCollection()
             this.disposables.push(this.contextDataCollector)
