+++ vscode/src/completions/inline-completion-item-provider.ts
@@ -209,7 +209,10 @@ export class InlineCompletionItemProvider
         )
         this.disposables.push(strategyFactory)
 
-        this.contextMixer = new ContextMixer(strategyFactory, true)
+        this.contextMixer = new ContextMixer({
+            strategyFactory,
+            dataCollectionEnabled: true,
+        })
         this.disposables.push(this.contextMixer)
 
         this.smartThrottleService = new SmartThrottleService()
