+++ vscode/src/supercompletions/supercompletion-provider.ts
@@ -5,7 +5,7 @@ import type { CodyStatusBar } from '../services/StatusBar'
 import { type Supercompletion, getSupercompletions } from './get-supercompletion'
 import { SupercompletionRenderer } from './renderer'
 
-const EDIT_HISTORY = 5 * 60 * 1000
+const EDIT_HISTORY_TIMEOUT = 5 * 60 * 1000
 const SUPERCOMPLETION_TIMEOUT = 2 * 1000
 
 export class SupercompletionProvider implements vscode.Disposable {
@@ -29,7 +29,7 @@ export class SupercompletionProvider implements vscode.Disposable {
         this.renderer = new SupercompletionRenderer()
         this.recentEditsRetriever = new RecentEditsRetriever(
             {
-                maxAgeMs: EDIT_HISTORY,
+                maxAgeMs: EDIT_HISTORY_TIMEOUT,
             },
             workspace
         )
